+++ lensless/eval/benchmark.py
@@ -106,17 +106,17 @@ def benchmark(
     dataloader = DataLoader(dataset, batch_size=batchsize, pin_memory=(device != "cpu"))
     model.reset()
     idx = 0
-    for lensless, lensed in tqdm(dataloader):
-        lensless = lensless.to(device)
-        lensed = lensed.to(device)
+    with torch.no_grad():
+        for lensless, lensed in tqdm(dataloader):
+            lensless = lensless.to(device)
+            lensed = lensed.to(device)
 
-        # add shot noise
-        if snr is not None:
-            for i in range(lensless.shape[0]):
-                lensless[i] = add_shot_noise(lensless[i], float(snr))
+            # add shot noise
+            if snr is not None:
+                for i in range(lensless.shape[0]):
+                    lensless[i] = add_shot_noise(lensless[i], float(snr))
 
-        # compute predictions
-        with torch.no_grad():
+            # compute predictions
             if batchsize == 1:
                 model.set_data(lensless)
                 prediction = model.apply(
@@ -126,113 +126,115 @@ def benchmark(
             else:
                 prediction = model.batch_call(lensless, **kwargs)
 
-        if unrolled_output_factor:
-            unrolled_out = prediction[-1]
-            prediction = prediction[0]
+            if unrolled_output_factor:
+                unrolled_out = prediction[-1]
+                prediction = prediction[0]
 
-        # Convert to [N*D, C, H, W] for torchmetrics
-        prediction = prediction.reshape(-1, *prediction.shape[-3:]).movedim(-1, -3)
-        lensed = lensed.reshape(-1, *lensed.shape[-3:]).movedim(-1, -3)
+            # Convert to [N*D, C, H, W] for torchmetrics
+            prediction = prediction.reshape(-1, *prediction.shape[-3:]).movedim(-1, -3)
+            lensed = lensed.reshape(-1, *lensed.shape[-3:]).movedim(-1, -3)
 
-        if crop is not None:
-            prediction = prediction[
-                ...,
-                crop["vertical"][0] : crop["vertical"][1],
-                crop["horizontal"][0] : crop["horizontal"][1],
-            ]
-            lensed = lensed[
-                ...,
-                crop["vertical"][0] : crop["vertical"][1],
-                crop["horizontal"][0] : crop["horizontal"][1],
-            ]
-
-        if save_idx is not None:
-            batch_idx = np.arange(idx, idx + batchsize)
-
-            for i, idx in enumerate(batch_idx):
-                if idx in save_idx:
-                    prediction_np = prediction.cpu().numpy()[i].squeeze()
-                    # switch to [H, W, C]
-                    prediction_np = np.moveaxis(prediction_np, 0, -1)
-                    save_image(prediction_np, fp=os.path.join(output_dir, f"{idx}.png"))
-
-        # normalization
-        prediction_max = torch.amax(prediction, dim=(-1, -2, -3), keepdim=True)
-        if torch.all(prediction_max != 0):
-            prediction = prediction / prediction_max
-        else:
-            print("Warning: prediction is zero")
-        lensed_max = torch.amax(lensed, dim=(1, 2, 3), keepdim=True)
-        lensed = lensed / lensed_max
-
-        # compute metrics
-        for metric in metrics:
-            if metric == "ReconstructionError":
-                metrics_values[metric].append(model.reconstruction_error().cpu().item())
-            else:
-                if "LPIPS" in metric:
-                    if prediction.shape[1] == 1:
-                        # LPIPS needs 3 channels
-                        metrics_values[metric].append(
-                            metrics[metric](
-                                prediction.repeat(1, 3, 1, 1), lensed.repeat(1, 3, 1, 1)
-                            )
-                            .cpu()
-                            .item()
-                        )
-                    else:
-                        metrics_values[metric].append(
-                            metrics[metric](prediction, lensed).cpu().item()
-                        )
-                else:
-                    metrics_values[metric].append(metrics[metric](prediction, lensed).cpu().item())
-
-        # compute metrics for unrolled output
-        if unrolled_output_factor:
-
-            # -- convert to CHW and remove depth
-            unrolled_out = unrolled_out.reshape(-1, *unrolled_out.shape[-3:]).movedim(-1, -3)
-
-            # -- extraction region of interest
             if crop is not None:
-                unrolled_out = unrolled_out[
+                prediction = prediction[
+                    ...,
+                    crop["vertical"][0] : crop["vertical"][1],
+                    crop["horizontal"][0] : crop["horizontal"][1],
+                ]
+                lensed = lensed[
                     ...,
                     crop["vertical"][0] : crop["vertical"][1],
                     crop["horizontal"][0] : crop["horizontal"][1],
                 ]
 
-            # -- normalization
-            unrolled_out_max = torch.amax(unrolled_out, dim=(-1, -2, -3), keepdim=True)
-            if torch.all(unrolled_out_max != 0):
-                unrolled_out = unrolled_out / unrolled_out_max
+            if save_idx is not None:
+                batch_idx = np.arange(idx, idx + batchsize)
 
-            # -- compute metrics
+                for i, idx in enumerate(batch_idx):
+                    if idx in save_idx:
+                        prediction_np = prediction.cpu().numpy()[i]
+                        # switch to [H, W, C] for saving
+                        prediction_np = np.moveaxis(prediction_np, 0, -1)
+                        save_image(prediction_np, fp=os.path.join(output_dir, f"{idx}.png"))
+
+            # normalization
+            prediction_max = torch.amax(prediction, dim=(-1, -2, -3), keepdim=True)
+            if torch.all(prediction_max != 0):
+                prediction = prediction / prediction_max
+            else:
+                print("Warning: prediction is zero")
+            lensed_max = torch.amax(lensed, dim=(1, 2, 3), keepdim=True)
+            lensed = lensed / lensed_max
+
+            # compute metrics
             for metric in metrics:
                 if metric == "ReconstructionError":
-                    # only have this for final output
-                    continue
+                    metrics_values[metric].append(model.reconstruction_error().cpu().item())
                 else:
                     if "LPIPS" in metric:
-                        if unrolled_out.shape[1] == 1:
+                        if prediction.shape[1] == 1:
                             # LPIPS needs 3 channels
                             metrics_values[metric].append(
                                 metrics[metric](
-                                    unrolled_out.repeat(1, 3, 1, 1), lensed.repeat(1, 3, 1, 1)
+                                    prediction.repeat(1, 3, 1, 1), lensed.repeat(1, 3, 1, 1)
                                 )
                                 .cpu()
                                 .item()
                             )
                         else:
+                            metrics_values[metric].append(
+                                metrics[metric](prediction, lensed).cpu().item()
+                            )
+                    else:
+                        metrics_values[metric].append(
+                            metrics[metric](prediction, lensed).cpu().item()
+                        )
+
+            # compute metrics for unrolled output
+            if unrolled_output_factor:
+
+                # -- convert to CHW and remove depth
+                unrolled_out = unrolled_out.reshape(-1, *unrolled_out.shape[-3:]).movedim(-1, -3)
+
+                # -- extraction region of interest
+                if crop is not None:
+                    unrolled_out = unrolled_out[
+                        ...,
+                        crop["vertical"][0] : crop["vertical"][1],
+                        crop["horizontal"][0] : crop["horizontal"][1],
+                    ]
+
+                # -- normalization
+                unrolled_out_max = torch.amax(unrolled_out, dim=(-1, -2, -3), keepdim=True)
+                if torch.all(unrolled_out_max != 0):
+                    unrolled_out = unrolled_out / unrolled_out_max
+
+                # -- compute metrics
+                for metric in metrics:
+                    if metric == "ReconstructionError":
+                        # only have this for final output
+                        continue
+                    else:
+                        if "LPIPS" in metric:
+                            if unrolled_out.shape[1] == 1:
+                                # LPIPS needs 3 channels
+                                metrics_values[metric].append(
+                                    metrics[metric](
+                                        unrolled_out.repeat(1, 3, 1, 1), lensed.repeat(1, 3, 1, 1)
+                                    )
+                                    .cpu()
+                                    .item()
+                                )
+                            else:
+                                metrics_values[metric + "_unrolled"].append(
+                                    metrics[metric](unrolled_out, lensed).cpu().item()
+                                )
+                        else:
                             metrics_values[metric + "_unrolled"].append(
                                 metrics[metric](unrolled_out, lensed).cpu().item()
                             )
-                    else:
-                        metrics_values[metric + "_unrolled"].append(
-                            metrics[metric](unrolled_out, lensed).cpu().item()
-                        )
 
-        model.reset()
-        idx += batchsize
+            model.reset()
+            idx += batchsize
 
     # average metrics
     if return_average: